KAFKA-8901; Extend consumer group command to use the new Admin API to delete consumer offsets (KIP-496)
It add support to delete offsets in the `kafka-consumer-group`.

*More detailed description of your change,
if necessary. The PR title and PR message become
the squashed commit message, so use a separate
comment to ping reviewers.*

*Summary of testing strategy (including rationale)
for the feature or bug fix. Unit and/or integration
tests are expected for any behaviour change and
system tests should be considered for larger changes.*

Author: David Jacot <djacot@confluent.io>

Reviewers: Gwen Shapira

Closes #7362 from dajac/KAFKA-8901-delete-offsets-command