MINOR: Fix producer timeouts in log divergence test (#7728)
This test was taking more than 5 minutes because the producer writes were timing out. The problem was that broker 100 was being shutdown before broker 101, which meant that the partition was still offline after broker 100 was restarted. The producer timeouts were not detected because the produce future was not checked. After the fix, test time drops to about 15s.

Reviewers: Ismael Juma <ismael@juma.me.uk>