KAFKA-9065; Fix endless loop when loading group/transaction metadata (#7554)
The current coordinator loading logic causes an infinite loop when there is a gap between the last record in the log and the log end offset. This is possible because of compaction if the active segment is empty. The patch fixes the problem by breaking from the loading loop when a read from the log returns no additional data.

Reviewers: Jason Gustafson <jason@confluent.io>